+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - improved `be.json`
 - improved `be.buffer`
 
+### fixed
+- now `be.equal` with zero negative return false
+
 ## [1.3.1] - 2017-07-04
 
 ### Added